Add BetterFileSearch
BetterFileSearch is an extremely tiny Wave webapp that enhances the file
search feature in explorer.

Usually, when searching your entire PC for a specific file with Windows
Explorer, it takes minutes until it fully loads all the results. This is
because it has to go through the hard drive cache and search index as it
searches.

BetterFileSearch starts up Windows Explorer with a standalone native
hidden module. It's the same as normal search, but it's much quicker.
Start typing and results come up almost instantly with no waiting.