Update Notepad++ for Wave v0.62
Wave v0.62 now supports the "Run" attribute, allowing for official sources to be used rather than having to resort to redistribution. This webapp file has been updated to use the official link and work with Wave despite the file to be run not being the same filename as the AppName.